update code.Can delete but it cannot update and show imge

+++ myshop-demo/src/main/java/com/myshop/controller/NewsController.java
@@ -8,12 +8,14 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.commons.CommonsMultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.myshop.model.News;
 import com.myshop.repository.NewsRepository;
@@ -49,10 +51,21 @@ public class NewsController {
 		return "redirect:/news";
 	}
 
-	/*@RequestMapping(value = "/news/delete/{id}", method = RequestMethod.GET)
-	public String delete(@PathVariable(value = "id") int id) {
-		News result = newsRepository.findByNews_id(id);
+	@RequestMapping(value = "/news/delete/{id}", method = RequestMethod.GET)
+	public String delete(@PathVariable("id") int id) {
+		News result = newsRepository.findByNewsid(id);
+		if (result == null) {
+			return "The id not exist";
+		}
 		newsRepository.delete(result);
 		return "redirect:/news";
-	}*/
+	}
+
+	@RequestMapping(value = "/news/update/{id}", method = RequestMethod.GET)
+	public void update(@PathVariable("id") int id, Model model) {
+		News result = newsRepository.findByNewsid(id);
+		model.addAttribute("newsupdate", result);
+	}
+	
+
 }
